requirements.txt
gitpython
google-cloud-storage